Install unp
Unp is a small perl script which makes extraction of any archive files a bit easier. It support several compressors and archiver programs, chooses the right one(s) automatically and extracts one or more files in one go. Unp is a very useful and time-saving utility, if you have to deal with archive files often.

https://help.ubuntu.com/community/Unp